code clean up and rename the main cmd
search.py -> build-index.py

move all web services related files to web/ (need more work to make it
run)
move all internal modules in cbindex/
