+++ cbindex/rare.py
