+++ requirements.txt
@@ -1,5 +1,3 @@
-tornado==5.0.2
 glob2
 elasticsearch
-docker
 fire
