+++ web/es-ik/ik/config/extra_single_word.dic
