+++ web/es-ik/ik/config/extra_single_word_full.dic
