+++ web/es-ik/ik/config/preposition.dic
