+++ web/es-ik/ik/config/quantifier.dic
