+++ web/requirements.txt
@@ -1,0 +1,4 @@
+tornado==5.0.2
+glob2
+elasticsearch
+docker
