+++ web/views/favicon.ico
